Move featured collection activities to Post scheduler
Move Add/Remove activities for sticky posts from Actor scheduler to
Post scheduler since the activities are about posts. Keep the Actor
scheduler simple with just a profile update for backwards compatibility.
