+++ includes/scheduler/class-actor.php
@@ -13,7 +13,6 @@ use Activitypub\Collection\Extra_Fields;
 use Activitypub\Collection\Outbox;
 
 use function Activitypub\add_to_outbox;
-use function Activitypub\get_post_id;
 use function Activitypub\is_user_type_disabled;
 
 /**
@@ -52,8 +51,8 @@ class Actor {
 
 		\add_action( 'transition_post_status', array( self::class, 'schedule_post_activity' ), 33, 3 );
 
-		\add_action( 'post_stuck', array( self::class, 'schedule_featured_add' ) );
-		\add_action( 'post_unstuck', array( self::class, 'schedule_featured_remove' ) );
+		\add_action( 'post_stuck', array( self::class, 'sticky_post_update' ) );
+		\add_action( 'post_unstuck', array( self::class, 'sticky_post_update' ) );
 
 		// User deletion handling.
 		\add_action( 'delete_user', array( self::class, 'schedule_user_delete' ), 10, 3 );
@@ -155,66 +154,18 @@ class Actor {
 	}
 
 	/**
-	 * Schedule an Add activity when a post is added to the featured collection.
+	 * Send a profile update when a post's sticky status changes.
 	 *
 	 * @param int $post_id The post ID.
 	 */
-	public static function schedule_featured_add( $post_id ) {
+	public static function sticky_post_update( $post_id ) {
 		$post = \get_post( $post_id );
 
 		if ( ! $post ) {
 			return;
 		}
 
-		// Also send the profile update for backwards compatibility.
 		self::schedule_profile_update( $post->post_author );
-		self::schedule_featured_update( $post, 'Add' );
-	}
-
-	/**
-	 * Schedule a Remove activity when a post is removed from the featured collection.
-	 *
-	 * @param int $post_id The post ID.
-	 */
-	public static function schedule_featured_remove( $post_id ) {
-		$post = \get_post( $post_id );
-
-		if ( ! $post ) {
-			return;
-		}
-
-		// Also send the profile update for backwards compatibility.
-		self::schedule_profile_update( $post->post_author );
-		self::schedule_featured_update( $post, 'Remove' );
-	}
-
-	/**
-	 * Schedule an Add or Remove activity for the featured collection.
-	 *
-	 * When a post's sticky status changes, this sends an Add or Remove activity
-	 * to notify followers about the change to the actor's featured collection.
-	 *
-	 * @see https://github.com/Automattic/wordpress-activitypub/issues/2795
-	 *
-	 * @param \WP_Post $post          The post object.
-	 * @param string   $activity_type The activity type ('Add' or 'Remove').
-	 */
-	private static function schedule_featured_update( $post, $activity_type ) {
-		// Get the actor.
-		$actor = Actors::get_by_id( $post->post_author );
-
-		if ( ! $actor || \is_wp_error( $actor ) ) {
-			return;
-		}
-
-		// Create the Add or Remove activity.
-		$activity = new Activity();
-		$activity->set_type( $activity_type );
-		$activity->set_actor( $actor->get_id() );
-		$activity->set_object( get_post_id( $post->ID ) );
-		$activity->set_target( $actor->get_featured() );
-
-		add_to_outbox( $activity, null, $post->post_author );
 	}
 
 	/**
