+++ includes/scheduler/class-post.php
@@ -7,7 +7,11 @@
 
 namespace Activitypub\Scheduler;
 
+use Activitypub\Activity\Activity;
+use Activitypub\Collection\Actors;
+
 use function Activitypub\add_to_outbox;
+use function Activitypub\get_post_id;
 use function Activitypub\get_wp_object_state;
 use function Activitypub\is_post_disabled;
 
@@ -29,6 +33,10 @@ class Post {
 		\add_action( 'add_attachment', array( self::class, 'transition_attachment_status' ) );
 		\add_action( 'edit_attachment', array( self::class, 'transition_attachment_status' ) );
 		\add_action( 'delete_attachment', array( self::class, 'transition_attachment_status' ) );
+
+		// Sticky post transitions (featured collection).
+		\add_action( 'post_stuck', array( self::class, 'schedule_featured_add' ) );
+		\add_action( 'post_unstuck', array( self::class, 'schedule_featured_remove' ) );
 	}
 
 	/**
@@ -165,5 +173,56 @@ class Post {
 		if ( true !== $scheduled ) {
 			add_to_outbox( $post, $type, $post->post_author );
 		}
+	}
+
+	/**
+	 * Schedule an Add activity when a post is added to the featured collection.
+	 *
+	 * @param int $post_id The post ID.
+	 */
+	public static function schedule_featured_add( $post_id ) {
+		self::schedule_featured_update( $post_id, 'Add' );
+	}
+
+	/**
+	 * Schedule a Remove activity when a post is removed from the featured collection.
+	 *
+	 * @param int $post_id The post ID.
+	 */
+	public static function schedule_featured_remove( $post_id ) {
+		self::schedule_featured_update( $post_id, 'Remove' );
+	}
+
+	/**
+	 * Schedule an Add or Remove activity for the featured collection.
+	 *
+	 * When a post's sticky status changes, this sends an Add or Remove activity
+	 * to notify followers about the change to the actor's featured collection.
+	 *
+	 * @see https://github.com/Automattic/wordpress-activitypub/issues/2795
+	 *
+	 * @param int    $post_id       The post ID.
+	 * @param string $activity_type The activity type ('Add' or 'Remove').
+	 */
+	private static function schedule_featured_update( $post_id, $activity_type ) {
+		$post = \get_post( $post_id );
+
+		if ( ! $post ) {
+			return;
+		}
+
+		$actor = Actors::get_by_id( $post->post_author );
+
+		if ( ! $actor || \is_wp_error( $actor ) ) {
+			return;
+		}
+
+		$activity = new Activity();
+		$activity->set_type( $activity_type );
+		$activity->set_actor( $actor->get_id() );
+		$activity->set_object( get_post_id( $post->ID ) );
+		$activity->set_target( $actor->get_featured() );
+
+		add_to_outbox( $activity, null, $post->post_author );
 	}
 }
