+++ tests/phpunit/tests/includes/scheduler/class-test-post.php
@@ -7,6 +7,7 @@
 
 namespace Activitypub\Tests\Scheduler;
 
+use Activitypub\Collection\Actors;
 use Activitypub\Scheduler\Post;
 
 /**
@@ -194,5 +195,97 @@ class Test_Post extends \Activitypub\Tests\ActivityPub_Outbox_TestCase {
 		$activitypub_id = \add_query_arg( 'p', $post_id, \home_url( '/' ) );
 
 		$this->assertNull( $this->get_latest_outbox_item( $activitypub_id ) );
+	}
+
+	/**
+	 * Test that sticking a post creates an Add activity for the featured collection.
+	 *
+	 * @covers ::schedule_featured_add
+	 * @covers ::schedule_featured_update
+	 */
+	public function test_sticky_post_creates_add_activity() {
+		$user_id = self::factory()->user->create( array( 'role' => 'author' ) );
+		$actor   = Actors::get_by_id( $user_id );
+
+		$post_id        = self::factory()->post->create( array( 'post_author' => $user_id ) );
+		$activitypub_id = \Activitypub\get_post_id( $post_id );
+
+		\stick_post( $post_id );
+
+		// Query for the Add activity by object ID and activity type.
+		$outbox_items = \get_posts(
+			array(
+				'post_type'   => 'ap_outbox',
+				'post_status' => 'pending',
+				'meta_query'  => array( // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_query
+					array(
+						'key'   => '_activitypub_object_id',
+						'value' => $activitypub_id,
+					),
+					array(
+						'key'   => '_activitypub_activity_type',
+						'value' => 'Add',
+					),
+				),
+			)
+		);
+
+		$this->assertCount( 1, $outbox_items );
+
+		$last_item = $outbox_items[0];
+
+		// Verify the activity content.
+		$activity = \json_decode( $last_item->post_content, true );
+		$this->assertEquals( 'Add', $activity['type'] );
+		$this->assertEquals( $actor->get_id(), $activity['actor'] );
+		$this->assertEquals( $activitypub_id, $activity['object'] );
+		$this->assertEquals( $actor->get_featured(), $activity['target'] );
+	}
+
+	/**
+	 * Test that unsticking a post creates a Remove activity for the featured collection.
+	 *
+	 * @covers ::schedule_featured_remove
+	 * @covers ::schedule_featured_update
+	 */
+	public function test_unsticky_post_creates_remove_activity() {
+		$user_id = self::factory()->user->create( array( 'role' => 'author' ) );
+		$actor   = Actors::get_by_id( $user_id );
+
+		$post_id        = self::factory()->post->create( array( 'post_author' => $user_id ) );
+		$activitypub_id = \Activitypub\get_post_id( $post_id );
+
+		// First stick, then unstick.
+		\stick_post( $post_id );
+		\unstick_post( $post_id );
+
+		// Query for the Remove activity by object ID and activity type.
+		$outbox_items = \get_posts(
+			array(
+				'post_type'   => 'ap_outbox',
+				'post_status' => 'pending',
+				'meta_query'  => array( // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_query
+					array(
+						'key'   => '_activitypub_object_id',
+						'value' => $activitypub_id,
+					),
+					array(
+						'key'   => '_activitypub_activity_type',
+						'value' => 'Remove',
+					),
+				),
+			)
+		);
+
+		$this->assertCount( 1, $outbox_items );
+
+		$last_item = $outbox_items[0];
+
+		// Verify the activity content.
+		$activity = \json_decode( $last_item->post_content, true );
+		$this->assertEquals( 'Remove', $activity['type'] );
+		$this->assertEquals( $actor->get_id(), $activity['actor'] );
+		$this->assertEquals( $activitypub_id, $activity['object'] );
+		$this->assertEquals( $actor->get_featured(), $activity['target'] );
 	}
 }
